feat(sdk-lib-mpc): move ecdsa hdtree from core
TICKET: WP-400
