+++ modules/sdk-lib-mpc/src/curves/index.ts
@@ -1,2 +1,4 @@
 export * from './baseCurve';
 export * from './secp256k1';
+export * from './hdTree';
+export * from './bip32';
